Fix CI build failures: widen types on all platforms, fix format strings
- Update PlatformGetFileSize, PlatformReadBytes, PlatformGetFileInfo
  signatures on osx, freebsd, nuttx, barrelfish, FreeRTOS, win to
  match the FILE_OFFSET header declarations from b53f009
- Fix format strings for size_t/ssize_t/FILE_OFFSET in reverse_proxy,
  websocket_handler, system_sockets, system_sockets_send
- Fix signed/unsigned comparison in mbedtls3 and openssl SSL send
- Enlarge etag buffer from 34 to 40 bytes for safety margin
- Add PY_CONTEXT_CHECK and bounds validation to Python file upload API
- Fix FILE_OFFSET type for PlatformGetFileInfo callers in py_plugin

+++ lib/ssl/mbedtls3.c
@@ -384,7 +384,7 @@ SOCKET_SEND_STATUS WebserverSSLSendNonBlocking(socket_info *s, const unsigned ch
 
 	if (ret > 0) {
 		if (bytes_send) *bytes_send = ret;
-		if (ret == len) {
+		if ((size_t)ret == len) {
 			return SOCKET_SEND_NO_MORE_DATA;
 		}
 		return SOCKET_SEND_SEND_BUFFER_FULL;
